templated CellConstantRefine works for 2d/3d; tested double and float with mblkcomm

+++ source/test/mblkcomm/CellMultiblockTest.C
@@ -13,7 +13,8 @@
 #include "SAMRAI/hier/PatchGeometry.h"
 #include "SAMRAI/hier/VariableDatabase.h"
 #include "SAMRAI/pdat/CellVariable.h"
-#include "SAMRAI/pdat/CellDoubleConstantRefine.h"
+//#include "SAMRAI/pdat/CellDoubleConstantRefine.h"
+#include "SAMRAI/pdat/CellConstantRefine.h"
 
 #include "MultiblockTester.h"
 
@@ -95,7 +96,7 @@ void CellMultiblockTest::registerVariables(
 
    for (int i = 0; i < nvars; ++i) {
       d_variables[i].reset(
-         new pdat::CellVariable<double>(d_dim,
+         new pdat::CellVariable<float>(d_dim,
             d_variable_src_name[i],
             d_variable_depth[i]));
 
@@ -126,14 +127,14 @@ void CellMultiblockTest::initializeDataOnPatch(
 
       for (int i = 0; i < static_cast<int>(d_variables.size()); ++i) {
 
-         std::shared_ptr<pdat::CellData<double> > cell_data(
-            SAMRAI_SHARED_PTR_CAST<pdat::CellData<double>, hier::PatchData>(
+         std::shared_ptr<pdat::CellData<float> > cell_data(
+            SAMRAI_SHARED_PTR_CAST<pdat::CellData<float>, hier::PatchData>(
                patch.getPatchData(d_variables[i], getDataContext())));
          TBOX_ASSERT(cell_data);
 
          hier::Box dbox = cell_data->getGhostBox();
 
-         cell_data->fillAll((double)block_id.getBlockValue());
+         cell_data->fillAll((float)block_id.getBlockValue());
 
       }
    }
@@ -180,8 +181,8 @@ void CellMultiblockTest::setPhysicalBoundaryConditions(
 
    for (int i = 0; i < static_cast<int>(d_variables.size()); ++i) {
 
-      std::shared_ptr<pdat::CellData<double> > cell_data(
-         SAMRAI_SHARED_PTR_CAST<pdat::CellData<double>, hier::PatchData>(
+      std::shared_ptr<pdat::CellData<float> > cell_data(
+         SAMRAI_SHARED_PTR_CAST<pdat::CellData<float>, hier::PatchData>(
             patch.getPatchData(d_variables[i], getDataContext())));
       TBOX_ASSERT(cell_data);
 
@@ -195,7 +196,7 @@ void CellMultiblockTest::setPhysicalBoundaryConditions(
                gcw_to_fill);
 
          if (!node_bdry[ni].getIsMultiblockSingularity()) {
-            cell_data->fillAll((double)(node_bdry[ni].getLocationIndex() + 100),
+            cell_data->fillAll((float)(node_bdry[ni].getLocationIndex() + 100),
                fill_box);
          }
       }
@@ -211,7 +212,7 @@ void CellMultiblockTest::setPhysicalBoundaryConditions(
                   gcw_to_fill);
 
             if (!edge_bdry[ei].getIsMultiblockSingularity()) {
-               cell_data->fillAll((double)(edge_bdry[ei].getLocationIndex()
+               cell_data->fillAll((float)(edge_bdry[ei].getLocationIndex()
                                            + 100),
                   fill_box);
             }
@@ -229,7 +230,7 @@ void CellMultiblockTest::setPhysicalBoundaryConditions(
                   gcw_to_fill);
 
             if (!face_bdry[fi].getIsMultiblockSingularity()) {
-               cell_data->fillAll((double)(face_bdry[fi].getLocationIndex()
+               cell_data->fillAll((float)(face_bdry[fi].getLocationIndex()
                                            + 100),
                   fill_box);
             }
@@ -256,8 +257,8 @@ void CellMultiblockTest::fillSingularityBoundaryConditions(
 
    for (int i = 0; i < static_cast<int>(d_variables.size()); ++i) {
 
-      std::shared_ptr<pdat::CellData<double> > cell_data(
-         SAMRAI_SHARED_PTR_CAST<pdat::CellData<double>, hier::PatchData>(
+      std::shared_ptr<pdat::CellData<float> > cell_data(
+         SAMRAI_SHARED_PTR_CAST<pdat::CellData<float>, hier::PatchData>(
             patch.getPatchData(d_variables[i], getDataContext())));
       TBOX_ASSERT(cell_data);
 
@@ -315,8 +316,8 @@ void CellMultiblockTest::fillSingularityBoundaryConditions(
                                                   patch_blk_id,
                                                   encon_blk_id);
 
-                  std::shared_ptr<pdat::CellData<double> > sing_data(
-                     SAMRAI_SHARED_PTR_CAST<pdat::CellData<double>, hier::PatchData>(
+                  std::shared_ptr<pdat::CellData<float> > sing_data(
+                     SAMRAI_SHARED_PTR_CAST<pdat::CellData<float>, hier::PatchData>(
                         encon_patch->getPatchData(
                            d_variables[i], getDataContext())));
                   TBOX_ASSERT(sing_data);
@@ -347,7 +348,7 @@ void CellMultiblockTest::fillSingularityBoundaryConditions(
          }
       } else {
          cell_data->fillAll(
-            (double)bbox.getLocationIndex() + 200.0, fill_box);
+            (float)bbox.getLocationIndex() + 200.0, fill_box);
       }
    }
 }
@@ -378,8 +379,8 @@ bool CellMultiblockTest::verifyResults(
    }
    hier::Box pbox = patch.getBox();
 
-   std::shared_ptr<pdat::CellData<double> > solution(
-      new pdat::CellData<double>(pbox, 1, tgcw));
+   std::shared_ptr<pdat::CellData<float> > solution(
+      new pdat::CellData<float>(pbox, 1, tgcw));
 
    hier::Box tbox(pbox);
    tbox.grow(tgcw);
@@ -398,10 +399,10 @@ bool CellMultiblockTest::verifyResults(
 
    for (int i = 0; i < static_cast<int>(d_variables.size()); ++i) {
 
-      double correct = (double)block_id.getBlockValue();
+      float correct = (float)block_id.getBlockValue();
 
-      std::shared_ptr<pdat::CellData<double> > cell_data(
-         SAMRAI_SHARED_PTR_CAST<pdat::CellData<double>, hier::PatchData>(
+      std::shared_ptr<pdat::CellData<float> > cell_data(
+         SAMRAI_SHARED_PTR_CAST<pdat::CellData<float>, hier::PatchData>(
             patch.getPatchData(d_variables[i], getDataContext())));
       TBOX_ASSERT(cell_data);
       int depth = cell_data->getDepth();
@@ -410,9 +411,9 @@ bool CellMultiblockTest::verifyResults(
       for (pdat::CellIterator ci(pdat::CellGeometry::begin(pbox));
            ci != ciend; ++ci) {
          for (int d = 0; d < depth; ++d) {
-            double result = (*cell_data)(*ci, d);
+            float result = (*cell_data)(*ci, d);
 
-            if (!tbox::MathUtilities<double>::equalEps(correct, result)) {
+            if (!tbox::MathUtilities<float>::equalEps(correct, result)) {
                tbox::perr << "Test FAILED: ...."
                           << " : cell index = " << *ci << std::endl;
                tbox::perr << "    Variable = " << d_variable_src_name[i]
@@ -448,9 +449,9 @@ bool CellMultiblockTest::verifyResults(
             for (pdat::CellIterator ci(pdat::CellGeometry::begin(*ng));
                  ci != ciend; ++ci) {
                for (int d = 0; d < depth; ++d) {
-                  double result = (*cell_data)(*ci, d);
+                  float result = (*cell_data)(*ci, d);
 
-                  if (!tbox::MathUtilities<double>::equalEps(correct,
+                  if (!tbox::MathUtilities<float>::equalEps(correct,
                          result)) {
                      tbox::perr << "Test FAILED: ...."
                                 << " : cell index = " << *ci << std::endl;
@@ -500,25 +501,25 @@ bool CellMultiblockTest::verifyResults(
 
                if (num_sing_neighbors == 0) {
 
-                  correct = (double)bdry[k].getLocationIndex() + 200.0;
+                  correct = (float)bdry[k].getLocationIndex() + 200.0;
 
                } else {
 
-                  correct /= (double)num_sing_neighbors;
+                  correct /= (float)num_sing_neighbors;
 
                }
 
             } else {
-               correct = (double)(bdry[k].getLocationIndex() + 100);
+               correct = (float)(bdry[k].getLocationIndex() + 100);
             }
 
             pdat::CellIterator ciend(pdat::CellGeometry::end(fill_box));
             for (pdat::CellIterator ci(pdat::CellGeometry::begin(fill_box));
                  ci != ciend; ++ci) {
                for (int d = 0; d < depth; ++d) {
-                  double result = (*cell_data)(*ci, d);
+                  float result = (*cell_data)(*ci, d);
 
-                  if (!tbox::MathUtilities<double>::equalEps(correct,
+                  if (!tbox::MathUtilities<float>::equalEps(correct,
                          result)) {
                      tbox::perr << "Test FAILED: ...."
                                 << " : cell index = " << *ci << std::endl;
